cast to int64

+++ src/main.go
@@ -335,7 +335,7 @@ func handleRootPost(w http.ResponseWriter, r *http.Request) {
 
 	// Announce successful payment via Nostr if enabled
 	var valueAfterFees = tokenValue - 2*mintFee
-	err = announceSuccessfulPayment(macAddress, valueAfterFees, durationSeconds)
+	err = announceSuccessfulPayment(macAddress, int64(valueAfterFees), durationSeconds)
 	if err != nil {
 	    log.Printf("Error announcing successful payment: %v", err)
 	}
